% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPower.R
\name{calcPower}
\alias{calcPower}
\title{...}
\usage{
calcPower(graph, mean, f = NULL, corr_sim, alpha = 0.025,
  n_sim = 10000)
}
\arguments{
\item{graph}{.}

\item{mean}{.}

\item{f}{Defaults to \code{NULL}.}

\item{corr_sim}{.}

\item{alpha}{Defaults to \code{0.025}.}

\item{n_sim}{Defaults to \code{1e4}.}
}
\value{
...
}
\description{
\code{calcPower}...
}
\examples{
# A basic example
basic <- calcPower(...)
}
\seealso{
\code{\link{g_mtp}}, ...
}
