% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/des_fixed.R
\name{des_fixed}
\alias{des_fixed}
\title{Calculate sample size for a fixed-sample two-arm randomised controlled trial}
\usage{
des_fixed(alpha = 0.05, beta = 0.2, delta = 0.5, sigma = 1,
  ratio = 1, two_sided = F)
}
\arguments{
\item{alpha}{A \code{\link{numeric}} giving the desired type-I error rate. It
must belong to (0,1). Defaults to \code{0.05}.}

\item{beta}{A \code{\link{numeric}} giving the desired type-II error rate. It
must belong to (0,1). Defaults to \code{0.2}.}

\item{delta}{A \code{\link{numeric}} giving the treatment effect to power
for. It must belong to (0,Inf). Defaults to \code{0.5}.}

\item{sigma}{A \code{\link{numeric}} giving the assumed value of the standard
deviation of the responses. It must belong to (0,Inf). Defaults to \code{1}.}

\item{ratio}{A \code{\link{numeric}} giving the allocation ratio to the
experimental arm relative to the control arm. It must belong to (0,Inf).
Defaults to \code{1}.}

\item{two_sided}{A \code{\link{logical}} variable specifying whether a one-
or two-sided null hypothesis should be assumed. Defaults to \code{F}.}
}
\value{
A \code{\link{list}} containing the following elements:
\itemize{
\item A \code{\link{numeric}} in the slot \code{$n} giving the sample size
required in the control arm.
\item A \code{\link{numeric}} in the slot \code{$N} giving the total required
sample size.
\item A \code{\link{numeric}} in the slot \code{$P_HG} giving the probability
of rejecting the null hypothesis when there is no treatment effect.
\item A \code{\link{numeric}} in the slot \code{$P_LFC} giving the
probability of rejecting the null hypothesis when the treatment effect is
\code{delta}.
\item A \code{\link{numeric}} in the slot \code{$u} giving the critical
rejection threshold for the test statistic.
}
}
\description{
\code{des_fixed} calculates the sample size required by a fixed-sample
(single-stage) two-arm randomised controlled trial that tests for
superiority, assuming a normally distributed outcome variable. It supports
unequal allocation to the two arms along with one- and two-sided null
hypotheses.
}
\examples{
# For the default parameters
default <- des_fixed()
}
