% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_rates.R
\name{error_rates}
\alias{error_rates}
\title{Error rates for normally distributed test statistics}
\usage{
error_rates(delta = numeric(5), correlation = 0,
  critical_value = 1.96)
}
\arguments{
\item{delta}{A \code{\link{numeric}} \code{\link{vector}} giving the
standardised means of the test statistics. It is assumed that the null
hypotheses are that each component of \code{delta} is less than or equal to 0
(so zero or negative \code{delta} values correspond to true nulls and
positive delta values correspond to false nulls). Its elements must be
finite. Defaults to \code{numeric(5)}.}

\item{correlation}{A \code{\link{numeric}} giving the correlation between the
test statistics. It must belong to [0,1]. Defaults to \code{0}.}

\item{critical_value}{A \code{\link{numeric}} giving the critical rejection
threshold such that when a test statistic takes a value above
\code{critical_value}, its corresponding null hypothesis is rejected.
Defaults to \code{1.96}.}
}
\value{
A \code{\link{list}} containing the following elements:
\itemize{
\item A \code{\link{numeric}} in the slot \code{$fdr} giving the estimated
FDR.
\item A \code{\link{numeric}} in the slot \code{$fwer} giving the FWER.
\item A \code{\link{numeric}} in the slot \code{$max_pher} giving the maximum
PHER.
}
}
\description{
\code{error_rates} finds the maximum per-hypothesis type I error-rate (PHER),
family-wise error-rate (FWER), and false-discovery rate (FDR) for normally
distributed test statistics with mean \code{delta}, variance 1, and a
specified level of correlation between each pair of test statistics (see
\code{correlation}).
}
\examples{
# For the default parameters
default <- error_rates()
}
\seealso{
\code{\link{power_of_test_corrections}},
\code{\link{test_corrections}}.
}
