% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gMCP.R
\name{gMCP}
\alias{gMCP}
\title{...}
\usage{
gMCP(graph, pvalues, alpha = 0.05, fweights = F)
}
\arguments{
\item{graph}{.}

\item{pvalues}{.}

\item{alpha}{Defaults to \code{0.05}.}

\item{fweights}{Defaults to \code{F}.}
}
\value{
...
}
\description{
\code{gMCP}...
}
\examples{
# A basic example
basic <- gMCP(...)
}
\seealso{
\code{\link{calcPower}}, ...
}
