% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtp_power.R
\name{mtp_power}
\alias{mtp_power}
\title{...}
\usage{
mtp_power(mu, sigma, alpha = 0.05, correction = "none",
  replicates = 1e+05)
}
\arguments{
\item{mu}{.}

\item{sigma}{.}

\item{alpha}{Defaults to \code{0.05}.}

\item{correction}{Defaults to \code{"none"}.}

\item{replicates}{Defaults to \code{1e5}.}
}
\value{
...
}
\description{
\code{mtp_power}...
}
\examples{
# A basic example
basic <- mtp_power()
}
\seealso{
...
}
