% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_of_test_corrections.R
\name{power_of_test_corrections}
\alias{power_of_test_corrections}
\title{Calculate power of multiple testing procedures}
\usage{
power_of_test_corrections(delta = numeric(5), correlation = 0,
  target_alpha = 0.05, replicates = 10000)
}
\arguments{
\item{delta}{A \code{\link{numeric}} \code{\link{vector}} giving the
standardised means of the test statistics. It is assumed that the null
hypotheses are that each component of \code{delta} is less than or equal to 0
(so zero or negative \code{delta} values correspond to true nulls and
positive delta values correspond to false nulls). Defaults to
\code{numeric(5)}.}

\item{correlation}{A \code{\link{numeric}} giving the correlation between
the first and second test statistics. Defaults to \code{0}.}

\item{target_alpha}{A \code{\link{numeric}} giving the target familywise
error-rate. It must belong to [-1,1]. Defaults to \code{0.05}.}

\item{replicates}{A \code{\link{numeric}} giving the number of replicate
simulations to use. It must belong to {2,3,...}. Defaults to \code{1e4}.}
}
\value{
A \code{\link{list}} containing the following elements:
\itemize{
\item A \code{\link{numeric}} \code{\link{vector}} in the slot
\code{$average_correct_rejections} giving the average number of correct
rejections for each of the three supported methods.
\item A \code{\link{numeric}} \code{\link{vector}} in the slot
\code{$average_incorrect_rejections} giving the average number of incorrect
rejections for each of the three supported methods.
}
}
\description{
\code{power_of_test_corrections} uses simulation to provide the expected
number of correct and incorrect rejections for each of three multiple testing
procedures: Benjamini-Hochberg, Bonferroni, and Holm.
}
\details{
Since it uses simulation, it may take a few seconds to run. If it takes too
long, then change \code{replicates} to a smaller number.

As in \code{\link{error_rates}}, \code{delta} is the vector of test statistic
means. However, in this case the correlation parameter (see
\code{correlation}) is only between the first and second test statistic
(to allow us to investigate negative values).
}
\examples{
# For the default parameters
default <- power_of_test_corrections()
}
\seealso{
\code{\link{error_rates}}, \code{\link{test_corrections}}.
}
