% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simConfint.R
\name{simConfint}
\alias{simConfint}
\title{...}
\usage{
simConfint(graph, pvalues, confint, alternative = c("less", "greater"),
  estimates, df, alpha = 0.05, mu = 0)
}
\arguments{
\item{graph}{.}

\item{pvalues}{.}

\item{confint}{.}

\item{alternative}{Defaults to \code{c("less", "greater")}.}

\item{estimates}{.}

\item{df}{.}

\item{alpha}{Defaults to \code{0.05}.}

\item{mu}{Defaults to \code{0}.}
}
\value{
...
}
\description{
\code{simConfInt}...
}
\examples{
# A basic example
basic <- simConfint()
}
\seealso{
...
}
