% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_corrections.R
\name{test_corrections}
\alias{test_corrections}
\title{Apply multiple testing procedures to p-values}
\usage{
test_corrections(pvalues, target_alpha = 0.05)
}
\arguments{
\item{pvalues}{A \code{\link{numeric}} \code{\link{vector}} giving the
p-values for each hypothesis. Its elements must all belong to (0,1].}

\item{target_alpha}{A \code{\link{numeric}} giving the target familywise
error-rate. It must belong to (0,1). Defaults to \code{0.05}.}
}
\value{
A \code{\link{list}} containing the following elements:
\itemize{
\item A \code{\link{numeric}} \code{\link{vector}} in the slot
\code{$benjamini_hochberg} giving the rejection results when applying
Benjamini-Hochberg.
\item A \code{\link{numeric}} \code{\link{vector}} in the slot
\code{$bonferroni} giving the rejection results when applying
Bonferroni.
\item A \code{\link{numeric}} \code{\link{vector}} in the slot
\code{$holm} giving the rejection results when applying Holm.
}
}
\description{
\code{test_corrections} allows you to apply Benjamini-Hochberg, Bonferroni,
and Holm, to a set of p-values. For each method, it returns a vector where
elements take value \code{\link[base]{TRUE}} if the corresponding null
hypothesis is rejected, and \code{\link[base]{FALSE}} otherwise.
}
\details{
Note that \code{\link[stats]{p.adjust}} implements more corrections.
}
\examples{
# A basic example with three hypotheses
three <- test_corrections(c(0.01, 0.02, 0.03))
}
\seealso{
\code{\link{error_rates}}, \code{\link{power_of_test_corrections}}.
}
